Switched to a JWT strategy for authenticating users
After successful login / register attempts, the client will be passed back a summary of the user object (no password) and the user's token. The expectation is to have the token stored in localStorage. The token will contain some commonly used data about the user, such as their username, email, avatar URL. This makes it possible for the client to decode the JWT and get this data rather than having to keep requesting it from the server